Correct window position on macOS (multi-monitor setup also)

+++ intern/ghost/GHOST_IWindow.hh
@@ -160,7 +160,7 @@ class GHOST_IWindow {
    * Returns the index of the display (monitor) the window is on.
    * \return The index of the display (for macOS, otherwise defaults to -1).
    */
-  virtual int16_t getDisplay() const = 0;
+  virtual int8_t getDisplay() const = 0;
 
   /**
    * Returns the state of the window (normal, minimized, maximized).
