+++ intern/ghost/intern/GHOST_WindowSDL.hh
@@ -108,7 +108,7 @@ class GHOST_WindowSDL : public GHOST_Window {
 
   void clientToScreen(int32_t inX, int32_t inY, int32_t &outX, int32_t &outY) const override;
 
-  int16_t getDisplay() const override;
+  int8_t getDisplay() const override;
 
   GHOST_TSuccess setState(GHOST_TWindowState state) override;
