+++ intern/ghost/intern/GHOST_WindowWayland.hh
@@ -127,7 +127,7 @@ class GHOST_WindowWayland : public GHOST_Window {
 
   GHOST_TSuccess setWindowCursorVisibility(bool visible) override;
 
-  int16_t getDisplay() const override;
+  int8_t getDisplay() const override;
 
   GHOST_TSuccess setState(GHOST_TWindowState state) override;
