+++ intern/ghost/intern/GHOST_WindowX11.hh
@@ -87,7 +87,7 @@ class GHOST_WindowX11 : public GHOST_Window {
 
   void clientToScreen(int32_t inX, int32_t inY, int32_t &outX, int32_t &outY) const override;
 
-  int16_t getDisplay() const override;
+  int8_t getDisplay() const override;
 
   GHOST_TWindowState getState() const override;
